refactor: useCardNumber에서 카드사 관련 로직 useCardCompany로 분리
