+++ hooks/src/hooks/useCardNumbers/useCardNumbers.ts
@@ -1,21 +1,13 @@
-import { useState, useMemo } from "react";
+import { useState } from "react";
 import { ValidationResult } from "./type";
 import { ErrorType } from "../../types/ErrorType";
-import { getCardCompany } from "../useCardCompany/getCardCompany";
+import { useCardCompany } from "../useCardCompany/useCardCompany";
+import { useCardFormatter } from "../useCardCompany/useCardCompany";
 
 const CARD_NUMBER_RULE = {
   INVALID_LENGTH_ERROR: "카드 번호는 4자리로 입력해 주세요.",
   MAX_LENGTH: 4,
 } as const;
-
-const FORMAT_PATTERNS: Record<string, number[]> = {
-  visa: [4, 4, 4, 4],
-  master: [4, 4, 4, 4],
-  diners: [4, 6, 4],
-  amex: [4, 6, 5],
-  unionPay: [4, 4, 4, 4],
-  default: [4, 4, 4, 4],
-};
 
 export default function useCardNumbers(): ValidationResult & {
   formatted: string;
@@ -45,8 +37,6 @@ export default function useCardNumbers(): ValidationResult & {
     setError(newError);
   };
 
-  const cardType = useMemo(() => getCardCompany(raw), [raw]);
-
   const numbers = Array.from({ length: 4 }, (_, i) =>
     raw.slice(
       i * CARD_NUMBER_RULE.MAX_LENGTH,
@@ -54,16 +44,9 @@ export default function useCardNumbers(): ValidationResult & {
     )
   );
 
-  const formatted = useMemo(() => {
-    const pattern = FORMAT_PATTERNS[cardType] || FORMAT_PATTERNS.default;
-    const parts: string[] = [];
-    let idx = 0;
-    for (const len of pattern) {
-      parts.push(raw.slice(idx, idx + len));
-      idx += len;
-    }
-    return parts.filter(Boolean).join(" ");
-  }, [raw, cardType]);
+  const cardType = useCardCompany(raw);
+
+  const formatted = useCardFormatter(raw, cardType);
 
   return { numbers, error, cardType, handleCardNumberChange, formatted };
 }
